1.APP配置信息同步调整；2.默认角色控制；3.车辆未绑定锁前轨迹信息判断报警状态空指针异常修复

+++ src/main/java/com/tipray/controller/RoleController.java
@@ -127,6 +127,12 @@ public class RoleController extends BaseAction {
                 | LogTypeConst.TYPE_DELETE | LogTypeConst.RESULT_DONE;
         StringBuffer description = new StringBuffer("删除角色：");
         try {
+            if (id < 21) {
+                logger.warn("角色【{}】是系统内置角色，不允许删除！", id);
+                type++;
+                description.append("失败，系统内置角色不允许删除！");
+                return Message.error("系统内置角色不允许删除！");
+            }
             roleService.deleteRolesById(id, app);
             description.append("成功！");
             return Message.success();
